+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ##### New
 * maximally lazy `AppSegment` -> `x3` speed up.
 * `[JpegMetaParser]` indexing support for segments: e.g. `parser['APP1']`.
+* `[JpegMetaParser]` iterator support: `for segment in parser: ...`.
 * `[JpegMetaParser]` support for `len()` function: `len(parser)` -> number of segments.
 * `BSD-3-Clause` license added.